requirements.txt
# Validated default stack: transformers==4.57.0 for Qwen3-VL-30B-A3B-Instruct.
# To run DeepSeek-VL2, switch transformers to 4.38.0 and install xformers separately.
accelerate
attrdict
chardet
cmake
datasets>=2.12.0
einops
hjson
huggingface-hub==0.36.2
imageio
json-repair
matplotlib
ninja
nltk
numpy
nvtx
omegaconf
openai
opencv-python
openpyxl
optimum>=1.17.1
packaging>=20.0
pandas
pillow
portalocker
protobuf
psutil
pylatexenc==2.10
python-dotenv
py-cpuinfo
pyarrow
qwen-vl-utils
requests
rich
safetensors
scikit-learn
scipy
sentencepiece
sty
sympy
tabulate
tiktoken
timeout-decorator
timm
tqdm
triton==3.4.0
transformers==4.57.0
typing_extensions
validators
xlsxwriter

# Additional dependencies required by eval_acc / VLMEvalKit.
dotenv
google-genai
gradio
ipdb
math-verify
openai-clip
setuptools
torchmetrics
